services/search/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sentence-transformers==2.7.0
faiss-cpu==1.7.4
yake==0.4.8
rank-bm25==0.2.2
numpy==1.24.3
spacy==3.7.2
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1-py3-none-any.whl
